cleanup verbose documentation, revert `AMPTestModel` location, refine new `_patch_cuda_is_available` context manager
