+++ src/lightning_lite/accelerators/cuda.py
@@ -82,13 +82,13 @@ def _get_all_available_cuda_gpus() -> List[int]:
 def _patch_cuda_is_available() -> Generator:
     """Context manager that safely patches :func:`torch.cuda.is_available` with its NVML-based version if
     possible."""
-    new_check = torch.cuda.device_count if _TORCH_GREATER_EQUAL_1_13 else _device_count_nvml
 
     if hasattr(torch._C, "_cuda_getDeviceCount") and _device_count_nvml() >= 0:
         # we can safely patch is_available if both torch has CUDA compiled and the NVML count is succeeding
         # otherwise, patching is_available could lead to attribute errors or infinite recursion
+        new_check = is_cuda_available
         orig_check = torch.cuda.is_available
-        torch.cuda.is_available = new_check  # type: ignore[assignment]
+        torch.cuda.is_available = new_check
         try:
             yield
         finally:
